Improved Main

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
-        long ticketPriceInKopecks = 945067;
-        long miles = ticketPriceInKopecks /100 /20;
+        int ticketPriceInKopecks = 945067;
+        int miles = ticketPriceInKopecks / 100 / 20;
 
         System.out.println(miles);
     }
